Refactored project strucure
