+++ Runtime/World/Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 90d5ac191293aa942a9445ebaa213414
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
