+++ Runtime/World/WalkableTilemap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6cab227788a8a4047a69999ac615087e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
