+++ Runtime/World/World.cs
@@ -1,3 +1,5 @@
+using SkelTech.RPEST.World.Objects;
+
 using System.Collections.Generic;
 
 using UnityEngine;
